refactor(frontend): split Header.tsx into components/Header/ subtree
Breaks the ~580-line Header.tsx into Header (orchestrator), DesktopHeader,
MobileHeader, MobileMenu, BreadcrumbBar, and NotificationItem, and relocates
the four components used exclusively by Header (AccountControls,
NotificationDropdown, OrganizationSwitcher, LanguageSelector) into the same
folder. Deduplicates the desktop/mobile notification bell+panel into one
shared NotificationDropdown. Also fixes the mobile bell's missing
aria-hidden, keys BreadcrumbBar items by href/label instead of index,
switches the org-list fetch to AbortController, extracts a repeated
Tailwind ternary in MobileMenu, and updates frontend/AGENTS.md's stale
file-tree entry to match.
