+++ frontend/AGENTS.md
@@ -12,9 +12,8 @@ src/
 ├── hooks/
 │   └── useApiClient.ts     React hook: returns api-client instance with token from useAuth()
 ├── components/
-│   ├── Header.tsx          Header with auth state (login/logout buttons, org switcher) + optional per-page breadcrumb/action bar (see `breadcrumb` prop)
+│   ├── Header/             Header with auth state (login/logout buttons, org switcher, notifications) + optional per-page breadcrumb/action bar (see `breadcrumb` prop) - split across Header.tsx (orchestrator), DesktopHeader.tsx, MobileHeader.tsx, MobileMenu.tsx, AccountControls.tsx, NotificationDropdown.tsx, NotificationItem.tsx, OrganizationSwitcher.tsx, LanguageSelector.tsx, BreadcrumbBar.tsx - all exclusive to Header, nothing here is imported outside this folder
 │   ├── Footer.tsx          Footer with links and social icons
-│   ├── OrganizationSwitcher.tsx        Dropdown to switch active org
 │   ├── CreateOrganizationModal.tsx     Modal form for org creation
 │   ├── VolunteerOpportunitiesList.tsx  Paginated list (size=10), filter bar, gated create button
 │   └── CreateVolunteerOpportunityModal.tsx  Modal form for opportunity creation
