+++ frontend/src/locales/de.json
@@ -19,7 +19,6 @@
       "myAchievements": "Meine Errungenschaften",
       "profileSettings": "Profileinstellungen",
       "administration": "Administration",
-      "organizationDashboard": "Organisations-Dashboard",
       "organization": "Organisation",
       "signOut": "Abmelden",
       "signIn": "Anmelden",
