+++ frontend/src/locales/en.json
@@ -19,7 +19,6 @@
       "myAchievements": "My Achievements",
       "profileSettings": "Profile Settings",
       "administration": "Administration",
-      "organizationDashboard": "Organization Dashboard",
       "organization": "Organization",
       "signOut": "Sign out",
       "signIn": "Sign in",
